Defining the BMA class
This commit contains four scripts.  These four scripts are the skeleton
of the BMA class.  They define the class, how to initialize it, and
also contain three get functions which will allow the user to access
some of the helpful components of a BMA class object.  Each file also
contains a help file and is documented in code as appropriate.